+++ web/src/settings_users.js
@@ -1,12 +1,10 @@
 import $ from "jquery";
 
-import render_settings_deactivation_user_modal from "../templates/confirm_dialog/confirm_deactivate_user.hbs";
 import render_settings_reactivation_bot_modal from "../templates/confirm_dialog/confirm_reactivate_bot.hbs";
 import render_settings_reactivation_user_modal from "../templates/confirm_dialog/confirm_reactivate_user.hbs";
 import render_admin_user_list from "../templates/settings/admin_user_list.hbs";
 
 import * as blueslip from "./blueslip";
-import * as bot_data from "./bot_data";
 import * as browser_history from "./browser_history";
 import * as channel from "./channel";
 import * as confirm_dialog from "./confirm_dialog";
@@ -23,6 +21,7 @@ import * as settings_bots from "./settings_bots";
 import * as settings_data from "./settings_data";
 import * as settings_panel_menu from "./settings_panel_menu";
 import * as timerender from "./timerender";
+import * as user_deactivation_ui from "./user_deactivation_ui";
 import * as user_profile from "./user_profile";
 
 const section = {
@@ -444,67 +443,6 @@ function start_data_load() {
     populate_users();
 }
 
-export function confirm_deactivation(user_id, handle_confirm, loading_spinner) {
-    // Knowing the number of invites requires making this request. If the request fails,
-    // we won't have the accurate number of invites. So, we don't show the modal if the
-    // request fails.
-    channel.get({
-        url: "/json/invites",
-        timeout: 10 * 1000,
-        success(data) {
-            let number_of_invites_by_user = 0;
-            for (const invite of data.invites) {
-                if (invite.invited_by_user_id === user_id) {
-                    number_of_invites_by_user = number_of_invites_by_user + 1;
-                }
-            }
-
-            const bots_owned_by_user = bot_data.get_all_bots_owned_by_user(user_id);
-            const user = people.get_by_user_id(user_id);
-            const realm_url = page_params.realm_uri;
-            const realm_name = page_params.realm_name;
-            const opts = {
-                username: user.full_name,
-                email: user.delivery_email,
-                bots_owned_by_user,
-                number_of_invites_by_user,
-                admin_email: people.my_current_email(),
-                realm_url,
-                realm_name,
-            };
-            const html_body = render_settings_deactivation_user_modal(opts);
-
-            function set_email_field_visibility() {
-                const $send_email_checkbox = $("#dialog_widget_modal").find(".send_email");
-                const $email_field = $("#dialog_widget_modal").find(".email_field");
-
-                $email_field.hide();
-                $send_email_checkbox.on("change", () => {
-                    if ($send_email_checkbox.is(":checked")) {
-                        $email_field.show();
-                    } else {
-                        $email_field.hide();
-                    }
-                });
-            }
-
-            dialog_widget.launch({
-                html_heading: $t_html(
-                    {defaultMessage: "Deactivate {name}?"},
-                    {name: user.full_name},
-                ),
-                help_link: "/help/deactivate-or-reactivate-a-user#deactivating-a-user",
-                html_body,
-                html_submit_button: $t_html({defaultMessage: "Deactivate"}),
-                id: "deactivate-user-modal",
-                on_click: handle_confirm,
-                post_render: set_email_field_visibility,
-                loading_spinner,
-            });
-        },
-    });
-}
-
 function handle_deactivation($tbody) {
     $tbody.on("click", ".deactivate", (e) => {
         // This click event must not get propagated to parent container otherwise the modal
@@ -528,7 +466,7 @@ function handle_deactivation($tbody) {
             dialog_widget.submit_api_request(channel.del, url, data);
         }
 
-        confirm_deactivation(user_id, handle_confirm, true);
+        user_deactivation_ui.confirm_deactivation(user_id, handle_confirm, true);
     });
 }
 
@@ -547,7 +485,7 @@ function handle_bot_deactivation($tbody) {
             dialog_widget.submit_api_request(channel.del, url);
         }
 
-        settings_bots.confirm_bot_deactivation(bot_id, handle_confirm, true);
+        user_deactivation_ui.confirm_bot_deactivation(bot_id, handle_confirm, true);
     });
 }
